requirements.txt
streamlit~=1.17.0
Pillow~=9.4.0